Generate simple mapping for un-ambiguous values
Currently, we only process one time for each execution, which indicates we only generate mock responses for each request per execution. For these responses, we only map the properties whose values are un-ambiguous, which implies we don't handle boolean properties, or polymorphic models.

The end result is a map whose key is the jsonpointer to each leaf property of the application model, and whose value is the property address of its swagger expanded model. Note that we are gonna improve the map value to make it more useful, e.g. to point to the exact schema that defines this property (probably giving a github link), together with the address refactored to be a jsonpointer also.